Move some more components to src/components/elements/reactions/scheme_tab
