+++ app/packs/src/components/elements/reactions/scheme_tab/MaterialCalculations.js
