+++ app/packs/src/components/elements/reactions/scheme_tab/MaterialGroup.js
@@ -2,8 +2,8 @@ import React from 'react';
 import PropTypes from 'prop-types';
 import { Button, Glyphicon, Tooltip, OverlayTrigger } from 'react-bootstrap';
 import VirtualizedSelect from 'react-virtualized-select';
-import Material from 'src/components/Material';
-import MaterialCalculations from 'src/components/MaterialCalculations';
+import Material from 'src/components/elements/reactions/scheme_tab/Material';
+import MaterialCalculations from 'src/components/elements/reactions/scheme_tab/MaterialCalculations';
 import ElementActions from 'src/stores/alt/actions/ElementActions';
 import MoleculesFetcher from 'src/fetchers/MoleculesFetcher';
 import Molecule from 'src/models/Molecule';
